Refactor Order data structure out of tinabot module
